# app/Http/Requests/StoreRoleRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *      @OA\Property(property="name",type="string"),
 *      @OA\Property(property="user_ids", type="array", @OA\Items(type="number")),
 * )
 */
class StoreRoleRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return false;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, mixed>
     */
    public function rules()
    {
        return [
            'name' => 'required|max:255',
            'user_ids' => 'required|array',
            'user_ids.*' => 'required|distinct|integer|min:1',
        ];
    }
}